backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-speech
python-dotenv
